Relocating of files to simplify lib structure
