--- lib/mongoid/geo/criteria.rb
@@ -1,6 +1,0 @@
-require 'mongoid/geo/criterion/complex'
-require 'mongoid/geo/criterion/inclusion'
-require 'mongoid/geo/criterion/outer_operator'
-require 'mongoid/geo/criterion/twin_operators'
-require 'mongoid/geo/inflections'
-require 'mongoid/geo/criteria_helpers'
