+++ lib/mongoid_geo/criteria_helpers.rb
