+++ lib/mongoid_geo/criterion/complex.rb
