+++ lib/mongoid_geo/criterion/outer_operator.rb
