+++ lib/mongoid_geo/criterion/twin_operators.rb
