+++ lib/mongoid_geo/haversine.rb
