+++ lib/mongoid_geo/index.rb
