+++ lib/mongoid_geo/inflections.rb
