+++ lib/mongoid_geo/model.rb
