+++ lib/mongoid_geo/models.rb
