+++ lib/mongoid_geo/near.rb
