+++ lib/mongoid_geo/point.rb
