+++ lib/mongoid_geo/unit.rb
